goauth2: Allow for empty redirect URLs.
The previous behavior was to use the value "oob" when no redirect URL
was specified, which prevented clients from actually having an empty
redirect URL.  This is a breaking changing for any clients that are
doing out-of-band code exchange, and were relying on this behavior.  The
fix for these clients is to manually specify a RedirectURL of "oob".

Fixes 17

R=adg
CC=golang-dev
https://codereview.appspot.com/9824044

Committer: Andrew Gerrand <adg@golang.org>