goauth2: use an expiry in the past instead of time.Now() in tests
The tests used time.Now() as an expired token expiry; time.Now() is not,
however, strictly speaking in the past. On a system with low time
resolution, or minor clock jitter, this will cause occasional test
failures. Use a time that is thoroughly in the past instead.
(I wish that Go had a clock package we could use for mock time.)

I did not add a comment in the code because this strikes me as
unlikely to accidentally regress, so a comment would just add noise.

Fixes issue 22.

R=adg, nightlyone
CC=golang-dev
https://codereview.appspot.com/21660043

Committer: Andrew Gerrand <adg@golang.org>